Move game details tab opening logic to Manage section
