+++ newIDE/app/src/GameDashboard/GamesList.js
@@ -1,15 +1,11 @@
 // @flow
 import { Trans, t } from '@lingui/macro';
 import * as React from 'react';
-import AuthenticatedUserContext from '../Profile/AuthenticatedUserContext';
-import { type Game, registerGame } from '../Utils/GDevelopServices/Game';
+import { type Game } from '../Utils/GDevelopServices/Game';
 import { GameCard } from './GameCard';
 import { ColumnStackLayout } from '../UI/Layout';
-import { GameRegistration } from './GameRegistration';
 import { type GameDetailsTab } from './GameDetails';
-import useAlertDialog from '../UI/Alert/useAlertDialog';
 import RouterContext from '../MainFrame/RouterContext';
-import { extractGDevelopApiErrorStatusAndCode } from '../Utils/GDevelopServices/Errors';
 import SearchBar from '../UI/SearchBar';
 import { useDebounce } from '../Utils/UseDebounce';
 import Fuse from 'fuse.js';
@@ -24,7 +20,6 @@ import { Column, Line, Spacer } from '../UI/Grid';
 import Text from '../UI/Text';
 import Paper from '../UI/Paper';
 import BackgroundText from '../UI/BackgroundText';
-import { getDefaultRegisterGamePropertiesFromProject } from '../Utils/UseGameAndBuildsManager';
 import UserEarnings from './Monetization/UserEarnings';
 
 const pageSize = 10;
@@ -70,15 +65,8 @@ type Props = {|
 
 const GamesList = ({ project, games, onRefreshGames, onOpenGameId }: Props) => {
   const {
-    routeArguments,
     addRouteArguments,
-    removeRouteArguments,
   } = React.useContext(RouterContext);
-  const { getAuthorizationHeader, profile } = React.useContext(
-    AuthenticatedUserContext
-  );
-  const { showAlert, showConfirmation } = useAlertDialog();
-  const [isGameRegistering, setIsGameRegistering] = React.useState(false);
   const [searchText, setSearchText] = React.useState<string>('');
   const [currentPage, setCurrentPage] = React.useState<number>(0);
 
@@ -99,94 +87,6 @@ const GamesList = ({ project, games, onRefreshGames, onOpenGameId }: Props) => {
       searchClient,
       currentPage,
     })
-  );
-
-  const onRegisterGame = React.useCallback(
-    async () => {
-      if (!profile || !project) return;
-
-      const { id } = profile;
-      try {
-        setIsGameRegistering(true);
-        await registerGame(
-          getAuthorizationHeader,
-          id,
-          getDefaultRegisterGamePropertiesFromProject({ project })
-        );
-        await onRefreshGames();
-      } catch (error) {
-        console.error('Unable to register the game.', error);
-        const extractedStatusAndCode = extractGDevelopApiErrorStatusAndCode(
-          error
-        );
-        if (extractedStatusAndCode && extractedStatusAndCode.status === 403) {
-          await showAlert({
-            title: t`Game already registered`,
-            message: t`The project currently opened is registered online but you don't have
-          access to it. Ask the original owner of the game to share it with you
-          to be able to manage it.`,
-          });
-        } else {
-          await showAlert({
-            title: t`Unable to register the game`,
-            message: t`An error happened while registering the game. Verify your internet connection
-          or retry later.`,
-          });
-        }
-      } finally {
-        setIsGameRegistering(false);
-      }
-    },
-    [getAuthorizationHeader, profile, project, showAlert, onRefreshGames]
-  );
-
-  React.useEffect(
-    () => {
-      // TODO: Move this logic in homepage.
-      const loadInitialGame = async () => {
-        // When games are loaded and we have an initial game id, try to open it.
-        const initialGameId = routeArguments['game-id'];
-        if (games && initialGameId) {
-          const game = games.find(game => game.id === initialGameId);
-          removeRouteArguments(['game-id']);
-          if (game) {
-            onOpenGameId(game.id);
-          } else {
-            // If the game is not in the list, then either
-            // - allow to register it, if it's the current project.
-            // - suggest to open the file before continuing, if it's not the current project.
-            if (project && project.getProjectUuid() === initialGameId) {
-              const answer = await showConfirmation({
-                title: t`Game not found`,
-                message: t`This project is not registered online. Register it now
-              to get access to leaderboards, player accounts, analytics and more!`,
-                confirmButtonLabel: t`Register`,
-              });
-              if (!answer) return;
-
-              await onRegisterGame();
-            } else {
-              await showAlert({
-                title: t`Game not found`,
-                message: t`The game you're trying to open is not registered online. Open the project
-              file, then register it before continuing.`,
-              });
-            }
-          }
-        }
-      };
-      loadInitialGame();
-    },
-    [
-      games,
-      routeArguments,
-      removeRouteArguments,
-      onRegisterGame,
-      showConfirmation,
-      showAlert,
-      project,
-      onOpenGameId,
-    ]
   );
 
   const getGamesToDisplayDebounced = useDebounce(
@@ -259,13 +159,6 @@ const GamesList = ({ project, games, onRefreshGames, onOpenGameId }: Props) => {
           <ChevronArrowRight />
         </IconButton>
       </Line>
-      {!isGameRegistering && (
-        <GameRegistration
-          project={project}
-          hideLoader
-          onGameRegistered={onRefreshGames}
-        />
-      )}
       {displayedGames.length > 0 ? (
         displayedGames.map(game => (
           <GameCard
