+++ newIDE/app/src/MainFrame/EditorContainers/HomePage/ManageSection/index.js
@@ -1,12 +1,15 @@
 // @flow
 import * as React from 'react';
-import { Trans } from '@lingui/macro';
+import { t, Trans } from '@lingui/macro';
 
 import SectionContainer, { SectionRow } from '../SectionContainer';
 import ErrorBoundary from '../../../../UI/ErrorBoundary';
 import AuthenticatedUserContext from '../../../../Profile/AuthenticatedUserContext';
 import GamesList from '../../../../GameDashboard/GamesList';
-import { type Game } from '../../../../Utils/GDevelopServices/Game';
+import {
+  registerGame,
+  type Game,
+} from '../../../../Utils/GDevelopServices/Game';
 import PlaceholderError from '../../../../UI/PlaceholderError';
 import PlaceholderLoader from '../../../../UI/PlaceholderLoader';
 import { Column, Line } from '../../../../UI/Grid';
@@ -23,6 +26,11 @@ import Window from '../../../../Utils/Window';
 import { getHelpLink } from '../../../../Utils/HelpLink';
 import { type GameDetailsTab } from '../../../../GameDashboard/GameDetails';
 import GameOverview from '../../../../GameOverview';
+import useAlertDialog from '../../../../UI/Alert/useAlertDialog';
+import RouterContext from '../../../RouterContext';
+import { getDefaultRegisterGamePropertiesFromProject } from '../../../../Utils/UseGameAndBuildsManager';
+import { extractGDevelopApiErrorStatusAndCode } from '../../../../Utils/GDevelopServices/Errors';
+import { GameRegistration } from '../../../../GameDashboard/GameRegistration';
 
 const publishingWikiArticle = getHelpLink('/publishing/');
 
@@ -60,7 +68,13 @@ const ManageSection = ({
     profile,
     onOpenCreateAccountDialog,
     onOpenLoginDialog,
+    getAuthorizationHeader,
   } = authenticatedUser;
+  const { showAlert, showConfirmation } = useAlertDialog();
+  const [isRegisteringGame, setIsRegisteringGame] = React.useState(false);
+  const { routeArguments, removeRouteArguments } = React.useContext(
+    RouterContext
+  );
 
   React.useEffect(
     () => {
@@ -81,6 +95,93 @@ const ManageSection = ({
     [profile, openedGame, setOpenedGameId]
   );
 
+  const onRegisterGame = React.useCallback(
+    async () => {
+      if (!profile || !project) return;
+
+      const { id } = profile;
+      try {
+        setIsRegisteringGame(true);
+        await registerGame(
+          getAuthorizationHeader,
+          id,
+          getDefaultRegisterGamePropertiesFromProject({ project })
+        );
+        await onRefreshGames();
+      } catch (error) {
+        console.error('Unable to register the game.', error);
+        const extractedStatusAndCode = extractGDevelopApiErrorStatusAndCode(
+          error
+        );
+        if (extractedStatusAndCode && extractedStatusAndCode.status === 403) {
+          await showAlert({
+            title: t`Game already registered`,
+            message: t`The project currently opened is registered online but you don't have
+          access to it. Ask the original owner of the game to share it with you
+          to be able to manage it.`,
+          });
+        } else {
+          await showAlert({
+            title: t`Unable to register the game`,
+            message: t`An error happened while registering the game. Verify your internet connection
+          or retry later.`,
+          });
+        }
+      } finally {
+        setIsRegisteringGame(false);
+      }
+    },
+    [getAuthorizationHeader, profile, project, showAlert, onRefreshGames]
+  );
+
+  React.useEffect(
+    () => {
+      const loadInitialGame = async () => {
+        // When games are loaded and we have an initial game id, try to open it.
+        const initialGameId = routeArguments['game-id'];
+        if (games && initialGameId) {
+          const game = games.find(game => game.id === initialGameId);
+          removeRouteArguments(['game-id']);
+          if (game) {
+            setOpenedGameId(game.id);
+          } else {
+            // If the game is not in the list, then either
+            // - allow to register it, if it's the current project.
+            // - suggest to open the file before continuing, if it's not the current project.
+            if (project && project.getProjectUuid() === initialGameId) {
+              const answer = await showConfirmation({
+                title: t`Game not found`,
+                message: t`This project is not registered online. Register it now
+              to get access to leaderboards, player accounts, analytics and more!`,
+                confirmButtonLabel: t`Register`,
+              });
+              if (!answer) return;
+
+              await onRegisterGame();
+            } else {
+              await showAlert({
+                title: t`Game not found`,
+                message: t`The game you're trying to open is not registered online. Open the project
+              file, then register it before continuing.`,
+              });
+            }
+          }
+        }
+      };
+      loadInitialGame();
+    },
+    [
+      games,
+      routeArguments,
+      removeRouteArguments,
+      onRegisterGame,
+      showConfirmation,
+      showAlert,
+      project,
+      setOpenedGameId,
+    ]
+  );
+
   const onBack = React.useCallback(
     () => {
       setCurrentTab('details');
@@ -92,6 +193,13 @@ const ManageSection = ({
   if (openedGame) {
     return (
       <SectionContainer flexBody>
+        {!isRegisteringGame && (
+          <GameRegistration
+            project={project}
+            hideLoader
+            onGameRegistered={onRefreshGames}
+          />
+        )}
         <GameOverview
           currentView={currentTab}
           setCurrentView={setCurrentTab}
